refactor(ipc): split daemon ipc contract into crate
