+++ crates/airc-cli/Cargo.toml
@@ -14,6 +14,7 @@ path = "src/main.rs"
 
 [dependencies]
 airc-core = { path = "../airc-core" }
+airc-ipc = { path = "../airc-ipc" }
 airc-protocol = { path = "../airc-protocol" }
 airc-transport = { path = "../airc-transport" }
 airc-daemon = { path = "../airc-daemon" }
