+++ crates/airc-daemon/Cargo.toml
@@ -5,11 +5,9 @@ edition = "2021"
 license = "MIT"
 publish = false
 
-# Long-running daemon runtime: IPC accept loop + typed Request/Response
-# dispatch + per-wire substrate state. The CLI (`airc-cli`) is a thin
-# client over this crate's `DaemonClient`; the daemon-mode subcommand
-# in the CLI calls `airc_daemon::run` to bring up the long-lived
-# process.
+# Long-running daemon runtime: IPC accept loop + per-wire substrate
+# state. The typed IPC contract/client live in `airc-ipc`; this crate
+# only hosts the daemon process and dispatches requests.
 #
 # Split out from `airc-cli` per the operating doc's needed crate
 # layout (grievance §5: CLI/Daemon Accumulating Policy → airc-cli
@@ -20,12 +18,12 @@ workspace = true
 
 [dependencies]
 airc-core = { path = "../airc-core" }
+airc-ipc = { path = "../airc-ipc" }
 airc-protocol = { path = "../airc-protocol" }
 airc-store = { path = "../airc-store" }
 airc-transport = { path = "../airc-transport" }
 
 async-trait = "0.1"
-ciborium = "0.2"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 base64 = "0.22"
