+++ crates/airc-ipc/src/response.rs
