+++ crates/airc-lib/src/airc.rs
@@ -25,7 +25,8 @@ use std::sync::atomic::{AtomicU64, Ordering};
 use std::sync::Arc;
 
 use airc_core::{ClientId, PeerId, TranscriptEvent};
-use airc_daemon::{peers_store, DaemonClient, LocalIdentity};
+use airc_daemon::{peers_store, LocalIdentity};
+use airc_ipc::DaemonClient;
 use airc_protocol::{PeerKeyRegistry, VerificationPolicy};
 use airc_store::{EventStore, SqliteEventStore};
 use airc_transport::LanTcpAdapter;
